Major update, but may be (partially) reverted (as not fully tested) in the future : - now list of current processes is dissociated from the listview -> the listview only get events from the ProcessProvider class which provides the list of processes (the list does not depend from the connection type : generic provider) - should have fixed all thread-safety issues concerning processes - show hidden processes is now broken (as the processList is NOT what provides the list of processes) - "deleted item in red"/"new item in green" time can now be specified manually - added try/catch/finally each time a semaphore is used - ParentProcessName property has been removed - list of services should use the same way of doing - list of processes in job should be modified a lot - list of tasks/jobs/connections should be also modified - CPU usage must be optimized - Does not work for remote (with server) monitoring anymore (frmServer to change)
The new design is more generic and coherent (no need to display anything like a lvProcess to have a valid list of current processes).

+++ YAPM/Common/Misc.vb
@@ -433,7 +433,9 @@ Namespace Common
                 If value Then
                     regKey.SetValue(Application.ProductName, Application.ExecutablePath)
                 Else
-                    regKey.DeleteValue(Application.ProductName)
+                    If regKey.GetValue(Application.ProductName) IsNot Nothing Then
+                        regKey.DeleteValue(Application.ProductName)
+                    End If
                 End If
             Catch ex As Exception
                 Misc.ShowDebugError(ex)
@@ -1159,7 +1161,9 @@ Namespace Common
             Debug.WriteLine("       Message : " & ex.Message)
             Debug.WriteLine("       Trace : " & ex.StackTrace)
             Debug.WriteLine("================================================================")
+#If RELEASE_MODE = 0 Then
             Beep()
+#End If
         End Sub
